Updating GPIO files to change pins
So far the board is accepting the programs, however currently there is some compilation issues. We need to ensure that everywhere it's anticipating the four LED pins for UC3C_EK board, I guess we need to change these to AVR32_PIN_PA019, that might not be the exact name for it, however I know at this point that underlying, it is simply defined as #define AVR32_PIN_PA019 19.